fix: add comment function
